Master branch reset -- READ DESCRIPTION
I have decided to start everything from scratch again, as I thought
evreything could be improved. But it was starting to get to a point
where I thought if I didn't reset now so I could start fresh I would
start running into huge problems and regret it big time.

All of the old code is on the branch called "old". Do NOT make commits
here, this branch is only here for historical pupouses and in-case
something goes wrong during the reset of the main github-branch.