Removed readUnsignedVarInt(DataInput in, boolean secure)
I decided to remove my version of readUnsignedVarInt because it turns out there already was a function that did that with a DataInput as well! And I know it works, because I tested it before deciding to removing my version of readUnsignedVarInt.

NOTE: This will causes some errors to spring up in the code. This is no big deal, just remove the boolean argument and the error will go away.